Share ConfigMessageTaskTest fixtures
